refactor: types/idea.ts, _types/idea.ts 통합
- _types/idea.ts로 파일 통합 및 경로 변경
- 기존에 types/idea.ts에서 정의한 Idea 삭제 후 _types/idea.ts를 확장하는 방식으로 수정 (Idea 타입들이 하나의 타입을 확장한 형태와 유사)
- IdeaCardProps도 기존 타입을 확장해서 사용하는 방식으로 수정
