+++ src/app/(with-sidebar)/issues/_components/idea-card/idea-card.styles.ts
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import { ISSUE_STATUS, VOTE_TYPE } from '@/constants/issue';
 import { theme } from '@/styles/theme';
 import { IssueStatus } from '@/app/(with-sidebar)/issues/_types';
-import { CardStatus } from '../../types/idea';
+import { CardStatus } from '@/app/(with-sidebar)/issues/_types/idea';
 
 export const Wrapper = styled.div`
   position: relative;
