+++ src/app/(with-sidebar)/issues/hooks/use-drag-and-drop.ts
@@ -1,6 +1,6 @@
 ﻿import { useState } from 'react';
 import { DragEndEvent, DragStartEvent, PointerSensor, useSensor, useSensors } from '@dnd-kit/core';
-import type { IdeaWithPosition, Position } from '@/app/(with-sidebar)/issues/types/idea';
+import type { IdeaWithPosition, Position } from '@/app/(with-sidebar)/issues/_types/idea';
 
 interface UseDragAndDropProps {
   ideas: IdeaWithPosition[];
