+++ src/app/(with-sidebar)/issues/hooks/use-ideas-with-temp.ts
@@ -1,5 +1,5 @@
 import { useCallback, useEffect, useMemo, useState } from 'react';
-import type { IdeaWithPosition } from '@/app/(with-sidebar)/issues/types/idea';
+import type { IdeaWithPosition } from '@/app/(with-sidebar)/issues/_types/idea';
 import { useIssueIdeaQuery } from '@/hooks/issues';
 
 const TEMP_IDEA_STORAGE_KEY = (issueId: string) => `temp-idea-${issueId}`;
