+++ src/hooks/issues/use-vote-mutation.ts
@@ -2,7 +2,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 import toast from 'react-hot-toast';
 import { useSseConnectionStore } from '@/app/(with-sidebar)/issues/store/use-sse-connection-store';
 import { postVote } from '@/lib/api/vote';
-import type { IdeaWithPosition } from '@/app/(with-sidebar)/issues/types/idea';
+import type { IdeaWithPosition } from '@/app/(with-sidebar)/issues/_types/idea';
 
 export const useVoteMutation = (issueId: string, ideaId: string) => {
   const queryClient = useQueryClient();
